backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
oracledb
python-dotenv
pydantic
alembic
requests